requirements.txt
requests
pandas
google-cloud-bigquery
pandas-gbq
google-auth